implement server-side WRQ; improve internals
This commit adds support for handling WRQ requests from clients to store
a file on the TFTP server. Similar to the RRQ implementation, this was
optimized to have a low footprint and to be quite fast.

Adding support also required several changes to the internals, which
also have been improved a lot. The process of receiving files and
sending files is each encapsulated to be usable for the client
implementation in the future (since the file transfer process doesn't
differ).

+++ src/tftp/options.rs
@@ -23,7 +23,7 @@ pub enum TftpOptionKind {
 pub enum TftpOption {
 	Blocksize(u16),
 	Timeout(Duration),
-	TransferSize(u16),
+	TransferSize(u32),
 	WindowSize(u16),
 }
 impl TftpOption {
@@ -87,7 +87,7 @@ impl OptionNegotiation {
 		}
 
 		if let Some(val) = raw_opts.get(consts::OPT_TRANSFERSIZE_IDENT) {
-			if let Ok(tf_size) = u16::from_str_radix(*val, 10) {
+			if let Ok(tf_size) = u32::from_str_radix(*val, 10) {
 				res.add_option(TftpOption::TransferSize(tf_size));
 			} else { return Err(()); }
 		}
@@ -116,7 +116,7 @@ impl OptionNegotiation {
 pub struct TftpOptions {
 	pub blocksize: u16,
 	pub timeout: Duration,
-	pub transfer_size: u16,
+	pub transfer_size: u32,
 	pub window_size: u16
 }
 impl TftpOptions {
